--- src/FieldLabel.js
@@ -1,2 +1,0 @@
-export { Filled as LabelFilled } from './FieldLabel/Filled.js'
-export { Regular as LabelRegular } from './FieldLabel/Regular.js'
