--- src/FieldLabel/constants.js
@@ -1,4 +1,0 @@
-export const shrinkedLabelFontSize = '12px'
-export const borderRadius = '4px'
-export const inputHeight = 56
-export const inputHeightDense = 44
