+++ src/Label/InputContainer.js
@@ -2,8 +2,6 @@ import React from 'react'
 import css from 'styled-jsx/css'
 import cx from 'classnames'
 
-import { inputHeight, inputHeightDense } from '../constants.js'
-
 const styles = css`
     div {
         height: 100%;
