+++ src/index.js
@@ -27,6 +27,7 @@ export { FieldSet } from './FieldSet.js'
 export { FileInput } from './FileInput.js'
 export { Field } from './Field.js'
 export { Help } from './Help.js'
+export { Input } from './Input.js'
 export { InputField } from './InputField.js'
 export { Legend } from './Legend.js'
 export { LinearLoader } from './LinearLoader.js'
@@ -35,6 +36,7 @@ export { MenuList } from './MenuList.js'
 export { MenuItem } from './MenuItem.js'
 export { Radio } from './Radio.js'
 export { ScreenCover } from './ScreenCover.js'
+export { Select } from './Select.js'
 export { SelectField } from './SelectField.js'
 export { Switch } from './Switch.js'
